feat: add comment and documentation to the app
