+++ src/main/scala/dotapir/config/Configs.scala
@@ -6,6 +6,7 @@ import zio.config.typesafe.TypesafeConfigProvider
 
 import com.typesafe.config.ConfigFactory
 object Configs:
+  // Create a ZLayer from a configuration located at `path`
   def makeConfigLayer[C](path: String)(using conf: Config[C], r: Tag[C]) =
     ZLayer(
       TypesafeConfigProvider
