--- src/main/scala/dotapir/http/endpoints/HealthEndpoints.scala
@@ -1,15 +1,0 @@
-package dotapir.http.endpoints
-
-import sttp.tapir.*
-import zio.*
-
-trait HealthEndpoint {
-  val healthEndpoint = endpoint
-    .tag("health")
-    .name("health")
-    .get
-    .in("health")
-    .out(stringBody)
-    .description("Health check")
-
-}
