+++ src/main/scala/dotapir/server/controllers/BaseController.scala
@@ -3,6 +3,7 @@ package dotapir.server.controllers
 import sttp.tapir.server.ServerEndpoint
 import zio.Task
 
+// Base controller trait with `route` value for all the controllers
 trait BaseController {
 
   val routes: List[ServerEndpoint[Any, Task]]
